requirements.txt
streamlit
torch
diffusers
transformers
accelerate
Pillow
